% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/artp_functions.R
\name{calculate_all_truncation_rtp}
\alias{calculate_all_truncation_rtp}
\title{Calculate rank truncated product for all truncation points.}
\usage{
calculate_all_truncation_rtp(rank_statistics)
}
\arguments{
\item{rank_statistics}{\code{vector} containing rank_statistics_to_test.}

\item{gene_set}{\code{vector} contianing genes in set to test.}

\item{gene_pvalues}{\code{data.frame} containing gene names and p-values.}

\item{pvalue_col}{\code{character} giving name of column containing gene
p-values.}

\item{randomize_gene_labels}{\code{boolean}. If \code{TRUE}, randomize gene
labels in \code{gene_pvalues} to perform permutations.}
}
\value{
\code{vector} of rtp for all truncation points in gene set.
}
\description{
Calculate rank truncated product for all truncation points.
}

