% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/artp_functions.R
\name{calculate_artp_perm}
\alias{calculate_artp_perm}
\title{Calculate empirical ARTP p-values using permutations.}
\usage{
calculate_artp_perm(gene_set, gene_pvalues, n_perm = 1000)
}
\arguments{
\item{gene_set}{\code{vector} contianing genes in set to test.}

\item{gene_pvalues}{\code{data.frame} containing gene names and p-values.}

\item{n_perm}{\code{integer} number of permutations to perform.}
}
\value{
\code{data.frame} containing empirical ARTP p-values.
}
\description{
Calculate empirical ARTP p-values using permutations.
}

