% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fishers_method_functions.R
\name{calculate_fishers_method_test_statistic}
\alias{calculate_fishers_method_test_statistic}
\title{Calculate Fishers method test statistic.}
\usage{
calculate_fishers_method_test_statistic(gene_rank_statistic)
}
\arguments{
\item{gene_rank_statistic}{\code{vector} containing rank statistic of genes
of interest.}
}
\description{
Calculate Fishers method test statistic.
}

