% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_rank_statistic.R
\name{calculate_rank_statistic}
\alias{calculate_rank_statistic}
\title{Calculate rank statistic}
\usage{
calculate_rank_statistic(gene_pvalues, pvalue_col = "Pvalue")
}
\arguments{
\item{gene_pvalues}{\code{data.frame} containing gene name and p-value}

\item{pvalue_col}{\code{character} or \code{integer} defining the column
containing p-values}
}
\value{
\code{data.frame} contining original \code{gene_pvalues} with an
additional column of the rank statistic
}
\description{
\deqn{r_i = rank of gene_i / K}
}
\details{
Where K = total number of genes
}

