% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/artp_functions.R
\name{calculate_rtp}
\alias{calculate_rtp}
\title{Calculate rank truncation product}
\usage{
calculate_rtp(num_genes, gene_rank_statistics)
}
\arguments{
\item{num_genes}{\code{integer} defining the truncation point to test}

\item{gene_rank_statistics}{\code{vector} containing gene rank statistics
calculated by \code{\link{calculate_rank_statistic}}}
}
\value{
\code{numeric} giving the RTP for the given truncation point
}
\description{
Calculate rank truncation product
}

