% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_permutation_pvalues.R
\name{calculate_simulated_pvalues}
\alias{calculate_simulated_pvalues}
\title{Calculate empirical p-values from permutation data.}
\usage{
calculate_simulated_pvalues(perm_data)
}
\arguments{
\item{perm_data}{\code{list} containing permutation data where length is the
number of simulations.}
}
\value{
\code{numeric} of the empirical p-value.
}
\description{
Calculate empirical p-values from permutation data.
}

