% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process_gene_sets.R
\name{check_gene_set_size}
\alias{check_gene_set_size}
\title{Check gene set size.}
\usage{
check_gene_set_size(gene_set, nMin, nMax)
}
\arguments{
\item{gene_set}{Gene set to test.}

\item{nMin}{Integer. Minimum number of genes.}

\item{nMax}{Integer. Maximum number of genes.}
}
\description{
Check gene set size to see if it is within the size boundaries.
}

