% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculate_permutation_pvalues.R
\name{count_a_less_than_b}
\alias{count_a_less_than_b}
\title{Count the number of elements in \code{a} are less than \code{b}}
\usage{
count_a_less_than_b(a, b)
}
\arguments{
\item{a}{thing that less}

\item{b}{thing that is more}
}
\value{
\code{integer} count of times \code{a} is less than \code{b}
}
\description{
Count the number of elements in \code{a} are less than \code{b}
}

