% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_genes_in_set.R
\name{extract_genes_in_set}
\alias{extract_genes_in_set}
\title{Extract genes in set from \code{data.frame} containing all gene p-values.}
\usage{
extract_genes_in_set(genes_in_set, gene_pvalues,
  randomize_gene_labels = FALSE)
}
\arguments{
\item{genes_in_set}{\code{vector} containing names of genes in set of
interest.}

\item{gene_pvalues}{\code{data.frame} containing gene names p-values and
rank_statistic.}

\item{randomize_gene_labels}{\code{boolean}. If \code{TRUE}, randomize gene
labels among all genes.}
}
\value{
\code{data.frame} containing subset of \code{gene_pvalues} found in
\code{genes_in_set}.
}
\description{
Extract genes in set from \code{data.frame} containing all gene p-values.
}

