% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process_gene_sets.R
\name{filter_gene_set_on_size}
\alias{filter_gene_set_on_size}
\title{Filter gene set on size.}
\usage{
filter_gene_set_on_size(all_gene_sets, ...)
}
\arguments{
\item{all_gene_sets}{List containing the gene sets.}

\item{...}{Parameters to pass to \code{check_gene_set_size}}
}
\description{
Filter gene set on size.
}

