% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fishers_method_functions.R
\name{fishers_method_permutation}
\alias{fishers_method_permutation}
\title{Calculate empirical p-value for Fisher's method.}
\usage{
fishers_method_permutation(gene_set, gene_pvalues, n_perm = 1000)
}
\arguments{
\item{gene_pvalues}{\code{data.frame} containing gene names and rank
statistic calculated by \code{\link{calculate_rank_statistic}}.}

\item{n_perm}{\code{integer} defining the number of permutations to perform}

\item{genes_in_set}{\code{vector} containing genes in the set of interest.}
}
\description{
\enumerate{
 \item{First calculate rank statistic for all genes being tested.}
 \item{Extract genes in set.}
 \item{Calculate fishers statistic for the observed data.}
 \item{Calculate fishers statistics for permuted data.}
 \item{Calculate empirical p-values using observed and permuted data.}
}
}

