% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/import_gmt_file.R
\name{format_gmt}
\alias{format_gmt}
\title{Format \code{.gmt} data.}
\usage{
format_gmt(gmt_data)
}
\arguments{
\item{gmt_data}{\code{.gmt} data as read using \code{scan}.}
}
\description{
Format \code{.gmt} data.
}

